requirements.txt
faiss-cpu
pdfplumber
pytesseract
Pillow
camelot-py
opencv-python
numpy
python-dotenv
requests
